+++ CHANGELOG.md
@@ -4110,3 +4110,4 @@
 - [2026-09-01 12:27] style(core): adjust design tokens, typography, and contrast in core
 - [2026-09-01 12:46] test(forms): implement widget integration tests for forms
 - [2026-09-01 13:49] test(theme): implement widget integration tests for theme
+- [2026-09-01 14:55] feat(test): implement responsive test view component